kafka_services/sync_consumer/requirements.txt
kafka-python==2.0.2
pydantic-settings==2.3.3
